rename to generated_assets.py
