+++ great_expectations/datasource/fluent/pandas_filesystem_datasource.pyi
@@ -15,7 +15,7 @@ from typing import (
 from great_expectations._docs_decorators import public_api as public_api
 from great_expectations.compatibility.typing_extensions import override
 from great_expectations.datasource.fluent import Sorter, _PandasFilePathDatasource
-from great_expectations.datasource.fluent.data_asset.path.pandas.dynamic_assets import (
+from great_expectations.datasource.fluent.data_asset.path.pandas.generated_assets import (
     CSVAsset,
     ExcelAsset,
     FeatherAsset,
