+++ tests/datasource/fluent/test_pandas_s3_datasource.py
@@ -13,7 +13,7 @@ import great_expectations.exceptions as ge_exceptions
 from great_expectations.compatibility import aws, pydantic
 from great_expectations.core.util import S3Url
 from great_expectations.datasource.fluent import PandasS3Datasource
-from great_expectations.datasource.fluent.data_asset.path.pandas.dynamic_assets import CSVAsset
+from great_expectations.datasource.fluent.data_asset.path.pandas.generated_assets import CSVAsset
 from great_expectations.datasource.fluent.data_asset.path.path_data_asset import (
     PathDataAsset,
 )
