+++ tests/integration/cloud/end_to_end/test_pandas_filesystem_datasource.py
@@ -20,7 +20,7 @@ if TYPE_CHECKING:
         DataAsset,
         PandasFilesystemDatasource,
     )
-    from great_expectations.datasource.fluent.data_asset.path.pandas.dynamic_assets import (
+    from great_expectations.datasource.fluent.data_asset.path.pandas.generated_assets import (
         CSVAsset,
         ParquetAsset,
     )
